+++ sentry/src/main/java/io/sentry/TracesSampler.java
@@ -1,5 +1,6 @@
 package io.sentry;
 
+import io.sentry.clientreport.DiscardReason;
 import io.sentry.util.Objects;
 import io.sentry.util.SampleRateUtils;
 import org.jetbrains.annotations.ApiStatus;
@@ -41,16 +42,21 @@ public final class TracesSampler {
     }
     Boolean profilesSampled = profilesSampleRate != null && sample(profilesSampleRate, sampleRand);
 
-    boolean tracesSamplerFailed = false;
     if (options.getTracesSampler() != null) {
-      Double samplerResult = null;
+      final Double samplerResult;
       try {
         samplerResult = options.getTracesSampler().sample(samplingContext);
       } catch (Throwable t) {
-        tracesSamplerFailed = true;
         options
             .getLogger()
             .log(SentryLevel.ERROR, "Error in the 'TracesSamplerCallback' callback.", t);
+        options
+            .getClientReportRecorder()
+            .recordLostEvent(DiscardReason.CALLBACK_ERROR, DataCategory.Transaction);
+        options
+            .getClientReportRecorder()
+            .recordLostEvent(DiscardReason.CALLBACK_ERROR, DataCategory.Span);
+        return new TracesSamplingDecision(false, null, sampleRand, false, null);
       }
       if (samplerResult != null) {
         return new TracesSamplingDecision(
@@ -75,8 +81,7 @@ public final class TracesSampler {
       return SampleRateUtils.backfilledSampleRand(parentSamplingDecision);
     }
 
-    final @Nullable Double tracesSampleRateFromOptions =
-        tracesSamplerFailed ? null : options.getTracesSampleRate();
+    final @Nullable Double tracesSampleRateFromOptions = options.getTracesSampleRate();
     final @NotNull Double downsampleFactor =
         Math.pow(2, options.getBackpressureMonitor().getDownsampleFactor());
     final @Nullable Double downsampledTracesSampleRate =